Split the scraper into different files
